refactor(Class): 클래스 구분
입출력 클래스 분리

분리
